requirements.txt
google-genai>=0.7.0
python-dotenv>=1.0.1
pytest>=8.0.0
pytest-cov>=5.0.0
coverage>=7.4.0
langgraph>=0.2.0
langchain-google-genai>=0.0.11
langchain-core>=0.2.0
